requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
riotwatcher==3.2.4
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
